Refactor `on_ready` method and remove `clear` function
In this commit, the `on_ready` method has been refactored with the bot's presence now being changed at the beginning of the method and the global variable `start_time` being set within this method. The `clear()` function call has been removed from this method. A new method `setup_hook` has been added which sets the global variables `owner` and `shutdown`, and fetches the user with the `OWNERID`. The `clear` function, used to clear the console, has been removed entirely. The bot's presence is no longer changed after the command tree is synced, this is now done in the `on_ready` method.