Update Docker build command in .gitlab-ci.yml.template
Updated the Docker build command in the `.gitlab-ci.yml.template` file to use the `DOCKER_USER` environment variable for dynamic Docker username assignment in the image tag. This change allows different users to build and push the Docker image to their respective Docker repositories. The image is tagged as `latest` and is built from the `./Basis` directory.