Restrict change_nickname to guilds and add nickname param
Added @discord.app_commands.guild_only decorator to ensure the
change_nickname command can only be used within a guild context.
Updated the command to include a new parameter 'nick' for the
bot's new nickname. The command now edits the bot's nickname
in the guild and sends a confirmation message to the user.