Update Dockerfile to ignore root user actions in pip
Modified the Dockerfile to add the `ignore` option to the `--root-user-action` flag in `pip` commands. This change affects the commands that upgrade `pip`, `setuptools`, and `wheel`, as well as the command that installs the requirements from `requirements.txt`. The update ensures that these commands execute without interruptions due to root user actions.